Asyptotic covariance of vec(beta) for a fitted envelope model.

This function computes the asymptotic covariance of vec(beta) and asymptotic standard error for elements in beta for a fitted envelope model.

Usage

asym.env(object, ...)

Arguments

object

A fitted env object.

Value

covMatrix

The asymptotic covariance of vec(beta). An rp by rp matrix. The covariance matrix returned is asymptotic. For the actual standard errors, multiply by 1/n.

asySE

The asymptotic standard error for elements in beta under the envelope model. An r by p matrix. The standard errors returned are asymptotic, for actual standard errors, multiply by 1/sqrt(n).

ratio

The asymptotic standard error ratio of the standard multivariate linear regression estimator over the envelope estimator, for each element in beta. An r by p matrix.

References

Cook, R. Dennis, Bing Li, and Francesca Chiaromonte. "Envelope models for parsimonious and efficient multivariate linear regression." Statist. Sinica 20 (2010): 927-1010.

See also