choose_env

Select the dimension of the envelope subspace using Bayesian information criterion, Akaike information criterion, and Likelihood ratio testing.

Usage

choose_env(X, Y)

Arguments

X

Predictors. An n by p matrix, p is the number of predictors. The predictors can be univariate or multivariate, discrete or continuous.

Y

Multivariate responses. An n by r matrix, r is the number of responses and n is the number of observations. The responses must be continuous variables.

dims

A vector. The dimensions to be chosen from. The default is 0 to r.

maxiter

Maximum number of iterations. Default value: 100.

ftol

Tolerance parameter for F. Default value: 1e-2.

verbose

Flag for print out model fitting process, logical 0 or 1. Default value: 0.

Value

result

Dimensions of the envelope subspace chosen by BIC, AIC, LRT(0.05), and LRT(0.01).

detail

-2*Loglik, BIC, AIC, degrees of freedom, and p-values for LRT test.

Details

Choose u for envelope model.

References

The codes are implemented based on the algorithm in Section 4.3 of Cook et al (2010).

Cook, R. Dennis, Bing Li, and Francesca Chiaromonte. "Envelope models for parsimonious and efficient multivariate linear regression." Statist. Sinica 20 (2010): 927-1010.

Examples

data(wheatprotein)
X <- wheatprotein[, 8]
Y <- wheatprotein[, 1:6]
choose_env(X, Y)
#> $result
#>       BIC       AIC LRT(0.05) LRT(0.01) 
#>         1         1         1         1 
#> 
#> $detail
#>   -2*Loglik      BIC      AIC df      p-value
#> 0  1761.188 1866.813 1815.188  6 3.161515e-11
#> 1  1701.595 1811.132 1757.595  5 9.484492e-01
#> 2  1701.001 1814.450 1759.001  4 9.665205e-01
#> 3  1700.590 1817.951 1760.590  3 9.842313e-01
#> 4  1700.452 1821.725 1762.452  2 9.906733e-01
#> 5  1700.434 1825.619 1764.434  1 9.771481e-01
#> 6  1700.433 1829.530 1766.433  0 1.000000e+00
#> 
#> attr(,"class")
#> [1] "choose_env"
m1 = env(X, Y, 1)
m1$beta
#>            [,1]
#> [1,] -1.0332435
#> [2,]  4.4729586
#> [3,]  3.6654398
#> [4,] -5.9952574
#> [5,]  0.5978577
#> [6,] -1.5923896