Fit the sparse envelope model.

Fit the sparse envelope model.

Usage

spenv(
  X,
  Y,
  u,
  lambda1 = NULL,
  lambda2 = NULL,
  ftol = 1e-04,
  maxiter = 100,
  verbose = 0,
  init = NULL
)

Arguments

X

Predictors. An n by p matrix, p is the number of predictors. The predictors can be univariate or multivariate, discrete or continuous.

Y

Multivariate responses. An n by r matrix, r is the number of responses and n is the number of observations. The responses must be continuous variables.

u

Dimension of the envelope. An integer between 0 and r.

lambda1

A user-supplied lambda sequence for the first stage.

lambda2

A user-supplied lambda sequence for the second stage.

ftol

Tolerance parameter for F. Default value: 1e-4.

maxiter

Maximum number of iterations. Default value: 100.

verbose

Flag for printing out the model fitting process, logical 0 or 1. Default value: 0.

init

The initial value for the envelope subspace. An r by u matrix. The default value is the one generated by the function initial_value.

Value

alpha

The estimated intercept in the envelope model. An r by 1 vector.

beta

The envelope estimator of the regression coefficients. An r by p matrix.

Gamma

The orthogonal basis of the envelope subspace. An r by u semi-orthogonal matrix.

Gamma0

The orthogonal basis of the complement of the envelope subspace. An r by r-u semi-orthogonal matrix.

eta

The coordinates of beta with respect to Gamma. A u by p matrix.

Sigma

The envelope estimator of the error covariance matrix. An r by r matrix.

Omega

The coordinates of Sigma with respect to Gamma. A u by u matrix.

Omega0

The coordinates of Sigma with respect to Gamma0. An r-u by r-u matrix.

loglik

The maximized log likelihood function. A real number.

paramNum

The number of parameters in the envelope model. A positive integer.

sigRes

The sample error covariance matrix. An r by r matrix.

sigY

The sample response covariance matrix. An r by r matrix.

sigX

The sample predictor covariance matrix. A p by p matrix.

n

The number of observations in the data. A positive integer.

r

The number of responses. A nonnegative integer.

u

Dimension of the envelope. An integer between 0 and r.

p

The number of predictors. A positive integer.

q

The number of active responses. A positive integer.

where_1

The active responses chosen by the sparse envelope model.

where_0

The inactive responses chosen by the sparse envelope model.

lambda

The tuning parameters chosen by BIC.

BIC_seq1

The sequence of BIC in the first stage.

BIC_seq2

The sequence of BIC in the second stage.

fit_time

The time costs for fitting the sparse envelope model.

Details

This function fits the sparse envelope model to the responses using maximum likelihood estimation. When the dimension of the envelope is between 1 and r-1, we implement the algorithm in Su Z, Zhu G, Chen X, and Yang Y (2016). When the dimension is r, the envelope model degenerates to standard multivariate linear regression. When the dimension is 0, it means that X and Y are uncorrelated, and the fitting is different.

References

The codes are implemented based on the algorithm in Su, Z., Zhu, G., Chen, X., and Yang, Y. (2016), Sparse Envelope Model: Efficient Estimation and Response Variable Selection in Multivariate Linear Regression. Biometrika. 103, 579-593.

See also

Author

Guangyu Zhu guangyuzhu@uri.edu

Examples

## Berkeley
data(Berkeley)
X = Berkeley$X
Y2 = Berkeley$Y[, c(1, 2, 21, 23)]
m1 = spenv(X, Y2, 1)
asym.spenv(m1)
#> $covMatrix
#>      [,1] [,2]      [,3]      [,4]
#> [1,]    0    0  0.000000  0.000000
#> [2,]    0    0  0.000000  0.000000
#> [3,]    0    0  3.367738 -3.054445
#> [4,]    0    0 -3.054445  3.323175
#> 
#> $asySE
#>          [,1]
#> [1,] 0.000000
#> [2,] 0.000000
#> [3,] 1.835140
#> [4,] 1.822958
#> 
#> $ratio
#>          [,1]
#> [1,]      Inf
#> [2,]      Inf
#> [3,] 8.382751
#> [4,] 8.501125
#> 
#> attr(,"class")
#> [1] "asym.spenv"