Fit the envelope based sparse partial least squares model.

Fit the envelope based sparse partial least squares model.

Usage

spxenv(
  X,
  Y,
  u,
  lambda1 = NULL,
  lambda2 = NULL,
  ftol = 0.01,
  maxiter = 100,
  eps2 = 1e-04,
  init = NULL,
  asy = 0,
  verbose = 0,
  init_method = 1,
  lambda_spice = 0.1
)

Arguments

X

Predictors. An n by p matrix, p is the number of predictors. The predictors can be univariate or multivariate, discrete or continuous.

Y

Multivariate responses. An n by r matrix, r is the number of responses and n is the number of observations. The responses must be continuous variables.

u

Dimension of the envelope. An integer between 0 and p.

lambda1

A user supplied lambda sequence for the first stage.

lambda2

A user supplied lambda sequence for the second stage.

ftol

Tolerance parameter for F. Default value: 1e-4.

maxiter

Maximum number of iterations. Default value: 100.

init

The initial value for the envelope subspace. An p by u matrix. Default value is the one generated by function initial_value.

verbose

Flag for print out model fitting process, logical 0 or 1. Default value: 0.

asym

Flag for calculating the asymptotic covariance of vec(beta) and asymptotic standard error for elements in beta for a fitted envelope model, logical 0 or 1. Default value: 0.

Value

mu

The estimated intercept in the envelope model. An r by 1 vector.

beta

The envelope estimator of the regression coefficients beta. A p by r matrix.

Gamma

The orthogonal basis of the envelope subspace. A p by u semi-orthogonal matrix.

Gamma0

The orthogonal basis of the complement of the envelope subspace. An p by p-u semi-orthogonal matrix.

eta

The coordinates of beta with respect to Gamma. A u by r matrix.

SigX

The envelope estimator of the covariance matrix of X. An p by p matrix.

Omega

The coordinates of Sigma with respect to Gamma. A u by u matrix.

Omega0

The coordinates of Sigma with respect to Gamma0. A p-u by p-u matrix.

sigYcX

The estimated conditional covariance matrix of Y given X. An r by r matrix.

loglik

The maximized log likelihood function. A real number.

paramNum

The number of parameters in the envelope model. A positive integer.

n

The number of observations in the data. A positive integer.

q

The number of active response variables. A positive integer.

iternum

The number of iterations. A positive integer.

lambda

The tuning parameters chosen by BIC.

where_1

The active responses chosen by the sparse envelope model.

where_0

The inactive responses chosen by the sparse envelope model.

BIC_seq1

The sequence of BIC in the first stage.

BIC_seq2

The sequence of BIC in the second stage.

covMatrix

The asymptotic covariance of vec(beta). An pr by pr matrix. The covariance matrix returned are asymptotic. For the actual standard errors, multiply by 1/n.

asySE

The asymptotic standard error for elements in beta under the envelope model. An p by r matrix. The standard errors returned are asymptotic, for actual standard errors, multiply by 1/sqrt(n).

ratio

The asymptotic standard error ratio of the standard multivariate linear regression estimator over the envelope estimator, for each element in beta. An r by p matrix.

fit_time

The time costs for fitting the envelope model.

Details

This function fits the sparse envelope model to the predictor using the maximum likelihood estimation. When the dimension of the envelope is between 1 and p-1, we implemented the algorithm in G. Zhu and Z. Su (2017). When the dimension is p, then the envelope model degenerates to the standard multivariate linear regression. When the dimension is 0, it means that X and Y are uncorrelated, and the fitting is different.

References

#' The codes are implemented based on the algorithm in Guangyu Zhu. Zhihua Su. "Envelope-based sparse partial least squares." Ann. Statist. 48 (1) 161 - 182, February 2020.

See also

Author

#' Guangyu Zhu and Zhihua Su