Wheat Protein Data

The protein content of ground wheat samples.

Usage

data(wheatprotein)

Format

A data frame with 50 observations on the following 8 variables:

V1 to V6

Measurements of the reflectance of NIR radiation by the wheat samples at 6 wavelengths in the range 1680-2310 nm. The measurements were made on the log(1/reflectance) scale.

V7

The protein content of each sample (in percent).

V8

Binary indicator, 0 for high protein content and 1 for low protein content. The cutoff point is if the protein content is smaller than 9.75.

Details

Wheat Protein Data.

The data are the result of an experiment to calibrate a near infrared reflectance (NIR) instrument for measuring the protein content of ground wheat samples. The protein content of each sample (in percent) was measured by the standard Kjeldahl method. In Fearn (1983), the problem is to find a linear combination of the measurements that predicts protein content. The estimated coefficients can then be entered into the instrument allowing the protein content of future samples to be read directly. The first 24 cases were used for calibration, and the last 26 samples were used for prediction.

References

Fearn, T. (1983). A misuse of ridge regression in the calibration of a near infrared reflectance instrument.

Examples

data(wheatprotein)
X <- wheatprotein[, 8]
Y <- wheatprotein[, 1:6]
choose_env(X, Y)
#> $result
#>       BIC       AIC LRT(0.05) LRT(0.01) 
#>         1         1         1         1 
#> 
#> $detail
#>   -2*Loglik      BIC      AIC df      p-value
#> 0  1761.188 1866.813 1815.188  6 3.161515e-11
#> 1  1701.595 1811.132 1757.595  5 9.484492e-01
#> 2  1701.001 1814.450 1759.001  4 9.665205e-01
#> 3  1700.590 1817.951 1760.590  3 9.842313e-01
#> 4  1700.452 1821.725 1762.452  2 9.906733e-01
#> 5  1700.434 1825.619 1764.434  1 9.771481e-01
#> 6  1700.433 1829.530 1766.433  0 1.000000e+00
#> 
#> attr(,"class")
#> [1] "choose_env"
m1 = env(X, Y, 1)
m1$beta
#>            [,1]
#> [1,] -1.0332435
#> [2,]  4.4729586
#> [3,]  3.6654398
#> [4,] -5.9952574
#> [5,]  0.5978577
#> [6,] -1.5923896